core-server/requirements.txt
# Core dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
PyYAML>=6.0

# AI dependencies
openai>=1.3.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0

# Voice processing
pyttsx3>=2.90
pydub>=0.25.0
gTTS>=2.4.0

# Database
mysql-connector-python>=8.1.0
